# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Glide",
    "description": "A calendar agent that reserves time to travel.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/me": {
      "get": {
        "tags": [
          "sample"
        ],
        "summary": "Get Me",
        "operationId": "get_me_api_me_get",
        "parameters": [
          {
            "name": "X-Glide-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Glide-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/session": {
      "post": {
        "tags": [
          "sample"
        ],
        "summary": "Create Demo Session",
        "operationId": "create_demo_session_api_demo_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/CreateDemoSessionRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/day": {
      "get": {
        "tags": [
          "sample"
        ],
        "summary": "Get Day",
        "operationId": "get_day_api_day_get",
        "parameters": [
          {
            "name": "requested_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Requested Date"
            }
          },
          {
            "name": "X-Glide-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Glide-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs": {
      "post": {
        "tags": [
          "sample"
        ],
        "summary": "Queue Run",
        "operationId": "queue_run_api_runs_post",
        "parameters": [
          {
            "name": "X-Glide-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Glide-Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RunRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunQueuedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "tags": [
          "sample"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "X-Glide-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Glide-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/activity": {
      "get": {
        "tags": [
          "sample"
        ],
        "summary": "Get Activity",
        "operationId": "get_activity_api_activity_get",
        "parameters": [
          {
            "name": "X-Glide-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Glide-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/events/{occurrence_id}": {
      "patch": {
        "tags": [
          "sample"
        ],
        "summary": "Edit Sample Event",
        "operationId": "edit_sample_event_api_demo_events__occurrence_id__patch",
        "parameters": [
          {
            "name": "occurrence_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Occurrence Id"
            }
          },
          {
            "name": "X-Glide-Session",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Glide-Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventEditRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEvent"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/reset": {
      "post": {
        "tags": [
          "sample"
        ],
        "summary": "Reset Sample",
        "operationId": "reset_sample_api_demo_reset_post",
        "parameters": [
          {
            "name": "X-Glide-Session",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Glide-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/decisions": {
      "get": {
        "tags": [
          "sample"
        ],
        "summary": "List Decisions",
        "operationId": "list_decisions_api_decisions_get",
        "parameters": [
          {
            "name": "X-Glide-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Glide-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/decisions/{decision_id}/resolve": {
      "post": {
        "tags": [
          "sample"
        ],
        "summary": "Resolve Decision",
        "operationId": "resolve_decision_api_decisions__decision_id__resolve_post",
        "parameters": [
          {
            "name": "decision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Decision Id"
            }
          },
          {
            "name": "X-Glide-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Glide-Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveDecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings": {
      "patch": {
        "tags": [
          "sample"
        ],
        "summary": "Patch Settings",
        "operationId": "patch_settings_api_settings_patch",
        "parameters": [
          {
            "name": "X-Glide-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Glide-Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pause": {
      "post": {
        "tags": [
          "sample"
        ],
        "summary": "Pause Automation",
        "operationId": "pause_automation_api_pause_post",
        "parameters": [
          {
            "name": "X-Glide-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Glide-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resume": {
      "post": {
        "tags": [
          "sample"
        ],
        "summary": "Resume Automation",
        "operationId": "resume_automation_api_resume_post",
        "parameters": [
          {
            "name": "X-Glide-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Glide-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateDemoSessionRequest": {
        "properties": {
          "day": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CreateDemoSessionRequest"
      },
...
      "DemoSessionResponse": {
        "properties": {
          "session": {
            "$ref": "#/components/schemas/SessionHandle"
          },
          "settings": {
            "$ref": "#/components/schemas/UserSettings"
          },
          "source_events": {
            "items": {
              "$ref": "#/components/schemas/CalendarEvent"
            },
            "type": "array",
            "title": "Source Events"
          },
          "sample_date": {
            "type": "string",
            "format": "date",
            "title": "Sample Date"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session",
          "settings",
          "source_events",
          "sample_date",
          "label"
        ],
        "title": "DemoSessionResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RunQueuedResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "status"
        ],
        "title": "RunQueuedResponse"
      },
      "RunRequest": {
        "properties": {
          "trigger": {
            "type": "string",
            "title": "Trigger",
            "default": "sample"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RunRequest"
      },
...
      "UserSettings": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "time_zone": {
            "type": "string",
            "title": "Time Zone"
          },
          "source_calendar_id": {
            "type": "string",
            "title": "Source Calendar Id"
          },
          "glide_calendar_id": {
            "type": "string",
            "title": "Glide Calendar Id"
          },
          "start_place": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlaceRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "earliest_departure": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Earliest Departure"
          },
          "mode": {
            "$ref": "#/components/schemas/TravelMode",
            "default": "driving"
          },
          "padding_minutes": {
            "type": "integer",
            "maximum": 60.0,
            "minimum": 0.0,
            "title": "Padding Minutes",
            "default": 10
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision",
            "default": 1
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user_id",
          "time_zone",
          "source_calendar_id",
          "glide_calendar_id"
        ],
        "title": "UserSettings"
      },
...
    }
  }
}